+++ src/days/day_02.rs
@@ -1,4 +1,4 @@
-pub(crate) fn a(input: &str) -> i64 {
+pub(crate) fn a(input: &str) -> u32 {
     let lines: Vec<Vec<&str>> = input.lines().map(|line| line.split(' ').collect()).collect();
 
     let mut depth = 0;
@@ -7,13 +7,13 @@ pub(crate) fn a(input: &str) -> i64 {
     for line in lines {
         match line[0] {
             "down" => {
-                depth += line[1].parse::<i64>().unwrap();
+                depth += line[1].parse::<u32>().unwrap();
             },
             "up" => {
-                depth -= line[1].parse::<i64>().unwrap();
+                depth -= line[1].parse::<u32>().unwrap();
             },
             "forward" => {
-                length += line[1].parse::<i64>().unwrap();
+                length += line[1].parse::<u32>().unwrap();
             }
             _ => {
                 panic!("Unknown instruction {}", line[0])
@@ -24,7 +24,7 @@ pub(crate) fn a(input: &str) -> i64 {
     return depth*length
 }
 
-pub(crate) fn b(input: &str) -> i64 {
+pub(crate) fn b(input: &str) -> u32 {
     let lines: Vec<Vec<&str>> = input.lines().map(|line| line.split(' ').collect()).collect();
 
     let mut depth = 0;
@@ -34,14 +34,14 @@ pub(crate) fn b(input: &str) -> i64 {
     for line in lines {
         match line[0] {
             "down" => {
-                aim += line[1].parse::<i64>().unwrap();
+                aim += line[1].parse::<u32>().unwrap();
             },
             "up" => {
-                aim -= line[1].parse::<i64>().unwrap();
+                aim -= line[1].parse::<u32>().unwrap();
             },
             "forward" => {
-                length += line[1].parse::<i64>().unwrap();
-                depth += aim * line[1].parse::<i64>().unwrap();
+                length += line[1].parse::<u32>().unwrap();
+                depth += aim * line[1].parse::<u32>().unwrap();
             }
             _ => {
                 panic!("Unknown instruction {}", line[0])
